% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_relationships.R
\docType{data}
\name{household_relationships}
\alias{household_relationships}
\title{Inter-household relationships}
\format{
Data frame with columns
\describe{
\item{hhid.x}{Source household ID}
\item{hhid.y}{Target household ID}
\item{village}{Village number}
\item{type}{Relationship type}
}
}
\source{
\href{https://doi.org/10.7910/21538/4POS1J}{Harvard Dataverse}
}
\usage{
data(household_relationships)
}
\description{
Data frame containing inter-household relationships.
}
\details{
All relationships are undirected.
Rows with \code{hhid.x} > \code{hhid.y} have been omitted for brevity.

There are no between-village relationships.

\code{type} values correspond to questions asked in Banerjee et al.'s (2013) surveys.
}
\references{
Banerjee, A., Chandrasekhar, A. G., Duflo, E., and Jackson, M. O. (2013).
The Diffusion of Microfinance.
\emph{Science} 341(6144):1236498.
\doi{10.1126/science.1236498}
}
\keyword{datasets}
