% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\docType{data}
\name{households}
\alias{households}
\title{Household attributes}
\format{
Data frame with columns
\describe{
\item{hhid}{Household ID}
\item{village}{Village number}
\item{caste}{Household caste}
}
}
\source{
\href{https://doi.org/10.7910/21538/4POS1J}{Harvard Dataverse}
}
\usage{
data(households)
}
\description{
Data frame containing household attributes.
}
\references{
Banerjee, A., Chandrasekhar, A. G., Duflo, E., and Jackson, M. O. (2013).
The Diffusion of Microfinance.
\emph{Science} 341(6144):1236498.
\doi{10.1126/science.1236498}
}
\keyword{datasets}
